update swal and fix some buggy

+++ src/sections/user/user-table-row.jsx
@@ -1,7 +1,9 @@
 import axios from 'axios';
+import Swal from 'sweetalert2';
 import { useState } from 'react';
 import PropTypes from 'prop-types';
 import { useNavigate } from 'react-router-dom';
+import withReactContent from 'sweetalert2-react-content';
 
 import Stack from '@mui/material/Stack';
 import Avatar from '@mui/material/Avatar';
@@ -34,6 +36,7 @@ export default function UserTableRow({
   const Navigate = useNavigate();
   const [open, setOpen] = useState(null);
   const router = useRouter();
+  const MySwal = withReactContent(Swal);
 
   const handleDeleteUser = async () => {
     if (!id) {
@@ -41,20 +44,39 @@ export default function UserTableRow({
       return;
     }
 
-    console.log('Deleting user with id:', id);
+    // Display a confirmation modal
+    MySwal.fire({
+      title: 'คุณแน่ใจหรือไม่?',
+      text: 'คุณจะไม่สามารถกู้คืนได้อีก!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        console.log('Deleting user with id:', id);
 
-    try {
-      const response = await axios.delete(`/api/employees/${id}`);
+        try {
+          const response = await axios.delete(`/api/employees/${id}`);
 
-      if (response.status !== 200) {
-        console.error('Error deleting user:', response.status, response.statusText);
-      } else {
-        console.log('Successfully deleted user');
-        router.reload('/user');
+          if (response.status !== 200) {
+            console.error('Error deleting user:', response.status, response.statusText);
+          } else {
+            console.log('Successfully deleted user');
+            // Show a success message
+            MySwal.fire({
+              icon: 'success',
+              title: 'User deleted successfully!',
+            });
+            // Reload the page or update the user list
+            router.reload('/user');
+          }
+        } catch (error) {
+          console.error('Error deleting user:', error);
+        }
       }
-    } catch (error) {
-      console.error('Error deleting user:', error);
-    }
+    });
   };
 
   const handleOpenMenu = (event) => {
